requirements.txt
streamlit==1.41.1
requests==2.32.3

# Required to load .env file
python-dotenv==1.0.1

# Required for oauth2 demno
streamlit-oauth==0.1.14

# Required for the chat bot demo
openai==1.59.7

# Reuired for charts demo
plotly==5.24.1

# Required for camera input demo
rembg==2.0.61
pillow==11.0.0
onnxruntime==1.20.1

# Required for process statistics demo
psutil==6.1.1